% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_land.R
\name{allocate_land}
\alias{allocate_land}
\title{Functions for creating individuals, households, and settlements}
\usage{
allocate_land(households)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Functions for creating individuals, households, and settlements
}
\examples{
allocate_time(households)
}
