% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_functions.R
\name{allocate_time}
\alias{allocate_time}
\title{Household level time allocation}
\usage{
allocate_time(households, precipitation, runoff, total_labor = 1,
  j = 0.2, k = 0.6, psi = 0.2, epsilon = 0.18)
}
\arguments{
\item{households}{Tibble of household agents.}

\item{epsilon}{}
}
\description{
Given knowledge of the irrigation system and simple heuristics
for relative returns to labor spent farming and maintaining
infrastructure, households solve a constrained optimization
problem [@david2015effect] to determine the proportion of
available time they should devote to each activity so as to
maximize their expected utility.
}
\examples{
allocate_time(households)
}
