% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_households.R
\name{bind_households}
\alias{bind_households}
\title{Title}
\usage{
bind_households(old_households, new_households)
}
\arguments{
\item{new_households}{}
}
\value{

}
\description{
Title
}
\examples{
old_households <- create_households(4);new_households <- create_households(5) \%>\% mutate(household = NA)

}
