% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_land_req.R
\name{calc_land_req}
\alias{calc_land_req}
\title{Functions for creating individuals, households, and settlements}
\usage{
calc_land_req(n_occupants, yield, fallow = T)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Functions for creating individuals, households, and settlements
}
\examples{
allocate_time(households)
}
