% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_functions.R
\name{create.households}
\alias{create.households}
\title{Functions for creating individuals, households, and settlements}
\usage{
create.households(x)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Functions for creating individuals, households, and settlements
}
\examples{
allocate_time(households)
}
