% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_functions.R
\name{create_world}
\alias{create_world}
\title{Functions for creating individuals, households, and settlements}
\usage{
create_world()
}
\arguments{
\item{n_settlements}{Number of starting settlement agents}

\item{n_households}{Number of starting household agents}

\item{n_individuals}{Number of individuals in each household, 2 by default.}

\item{wheat_req}{kg of wheat to feed a person for 1 year}
}
\description{
Functions for creating individuals, households, and settlements
}
\examples{
create_settlement(4)
create_household(4)
create_occupant(4)
}
