% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eat.R
\name{eat}
\alias{eat}
\title{Consume food for storage}
\usage{
eat(households)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Consume food for storage
}
\examples{
allocate_time(households)
}
