% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_dynamics.R
\name{environmental_dynamics}
\alias{environmental_dynamics}
\title{Environmental dynamics}
\usage{
environmental_dynamics(settlements)
}
\arguments{
\item{settlements}{Tibble of settlement agents.}
}
\description{
This function takes the master dataframe and calculates the
per patch climatic yields from per patch precipitation and irrigation
}
\examples{
allocate_time(households)
}
