% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm.R
\name{farm}
\alias{farm}
\title{Functions for creating individuals, households, and settlements}
\usage{
farm(households)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Functions for creating individuals, households, and settlements
}
\examples{
allocate_time(households)
}
