% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fission.R
\name{fission}
\alias{fission}
\title{Household fusioning}
\usage{
fission(households, fission_rate = 0.2)
}
\arguments{
\item{households}{Tibble of household agents containg nested tibbles of individual agents}
}
\description{
This function allows household agents to split when they get too big.
}
\examples{
reproduce(households)
die(households)
}
