% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_dynamics.R
\name{household_dynamics}
\alias{household_dynamics}
\title{Household Dynamics}
\usage{
household_dynamics(settlements)
}
\arguments{
\item{settlements}{Tibble of settlement agents.}
}
\description{
This function takes a settlement (tibble of households), and makes the households do what they need to do.
}
\examples{
allocate_time(households)
}
