% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure_performance.R
\name{infrastructure_performance}
\alias{infrastructure_performance}
\title{Infrastructure performance}
\usage{
infrastructure_performance(maintainance_labor, psi = 0.2,
  epsilon = 0.18, max_irrigation = 1)
}
\arguments{
\item{maintainance_labor}{The proportion of total labor allocated to maintaining infrastructure.}

\item{psi}{The proportion of a household's labor needed to keep irrigation infrastructure at half capacity, defaults to 0.2.}

\item{epsilon}{The scalability of irrigation infrastructure, defaults to 0.18.}

\item{max_irrigation}{Maximum irrigation, defaults to 1.}
}
\description{
This function calculates the performance of irrigation
infrastructure, given an amount of maintainance labor and
parameters controling the scalability of the infrastructure.
The performance of infrastructure is a piecewise linear
function of labor inputs. Two parameters
\eqn{\psi} and \eqn{\epsilon} determine how much labor is required to
keep irrigation infrastructure working at maximum capacity.
By default, \eqn{\psi \approx \epsilon} to make the
infrastructure scalable, that is the agents can spend more
or less time maintaining infrastructure and still be assured
of at least some water. This equation is derived from [@david2015effect].
}
\examples{
infrastructure_performance(maintainance_labor = 0.5)
}
