% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_cultivable_land.R
\name{max_cultivable_land}
\alias{max_cultivable_land}
\title{Functions for creating individuals, households, and settlements}
\usage{
max_cultivable_land(laborers, farming_labor, available_area, fallow = T,
  type = "asymptote")
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Functions for creating individuals, households, and settlements
}
\examples{
allocate_time(households)
}
