% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_food.R
\name{produce_food}
\alias{produce_food}
\title{Calcualte yields and whatnot}
\usage{
produce_food(households)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Here households calculate how much land they need, pull the
crop yield from the environment, remember the yield, and
determine their harvests by multiplying the yield by the amount
of land they have (also removing some of the crop to save as
seed for next year).
}
\examples{

}
