% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation
We run the actual simulation using the *accumulate* function
from purrr. Under the hood its doing the same thing as a for loop,
but saving all the intervening steps for us to make plotting
and analysis easier later.}
\usage{
run_simulation(input_data, nsim, replicates = 1, intermediate = TRUE)
}
\arguments{
\item{replicates}{}
}
\description{
Run the simulation
We run the actual simulation using the *accumulate* function
from purrr. Under the hood its doing the same thing as a for loop,
but saving all the intervening steps for us to make plotting
and analysis easier later.
}
\examples{


have an option for 'save intermediate files', then make decide whether to use reduce or accumulate

}
