% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settlement_dynamics.R
\name{settlement_dynamics}
\alias{settlement_dynamics}
\title{Settlement dynamics}
\usage{
settlement_dynamics(settlements)
}
\arguments{
\item{net}{}
}
\value{

}
\description{
Settlement dynamics
}
