% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_firewood.R
\name{stage_to_biomass}
\alias{stage_to_biomass}
\title{Functions for converting vegetation succession stage to woody biomass (per m2) and back again.}
\usage{
stage_to_biomass(stage)
}
\arguments{
\item{vegetation}{}
}
\value{

}
\description{
Functions for converting vegetation succession stage to woody biomass (per m2) and back again.
}
