% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_to.R
\name{zoom_in2}
\alias{zoom_in2}
\title{Zoom functions}
\usage{
zoom_in2(x)
}
\arguments{
\item{FUN}{}
}
\value{

}
\description{
These are  internal convencience functions used by the *_dynamics commands
that allow functions meant for one level to be called from another level.
zoom_in takes a settlement or household tibble and produces as household or
individuals tibble, respectively. zoom_out does the opposite.
}
\examples{

}
