% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_time.R
\name{allocate_time}
\alias{allocate_time}
\title{Household level time allocation}
\usage{
allocate_time(households)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Given knowledge of the irrigation system and simple heuristics
for relative returns to labor farming and maintaining
infrastructure, households solve a constrained optimization
problem [@david2015effect] to determine of the proportion of
available time they should devote to each activity so as to
maximize their expected utility.
}
\examples{
allocate_time(households)
}
