% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm.R
\name{farm}
\alias{farm}
\title{Functions for calculating climatic yields and fertility impact on yields}
\usage{
farm(households)
}
\arguments{
\item{households}{Tibble of household agents.}
}
\description{
Crop yields are determined by rainfall and soil fertility.
The climatic potential yield (determined by precipitation) is
calculated separately from successive local yield reduction
factors. This allows the climatic potential yield to be easily
substituted by potential yields from the crop model component
of an Earth System Model in future simulations.
}
\examples{

}
